+++ alteia/apis/provider.py
@@ -3,25 +3,29 @@ import json
 from alteia.core.connection.connection import Connection
 from alteia.core.utils.utils import sanitize_dict
 
+DEFAULT_API_TIMEOUT = 30.0  # value in seconds
+
 
 class Provider(object):
     _root_path = ''
+    api_timeout = DEFAULT_API_TIMEOUT
 
     def __init__(self, connection: Connection):
         self._connection = connection
 
-    def get(self, path, *, preload_content=True, as_json=True):
+    def get(self, path, *, preload_content=True, as_json=True, timeout=None):
         headers = {'Cache-Control': 'no-cache'}
         full_path = '{root}/{path}'.format(root=self._root_path, path=path)
         content = self._connection.get(path=full_path,
                                        headers=headers,
+                                       timeout=timeout or self.api_timeout,
                                        as_json=as_json,
                                        preload_content=preload_content)
 
         return content
 
     def post(self, path, data, *, sanitize=False, serialize=True,
-             preload_content=True, as_json=True):
+             preload_content=True, as_json=True, timeout=None):
         """Post the given data.
 
         Args:
@@ -37,6 +41,8 @@ class Provider(object):
             preload_content: Whether to preload the response content.
 
             as_json: Whether to deserialize the response body from JSON.
+
+            timeout: Timeout in seconds for API call
 
         Returns:
             Response body eventually deserialized.
@@ -57,12 +63,13 @@ class Provider(object):
         content = self._connection.post(path=full_path,
                                         headers=headers,
                                         data=data,
+                                        timeout=timeout or self.api_timeout,
                                         as_json=as_json,
                                         preload_content=preload_content)
         return content
 
     def put(self, path, data, *, sanitize=True, serialize=True,
-            preload_content=True, as_json=True):
+            preload_content=True, as_json=True, timeout=None):
         """Put the given data.
 
         Args:
@@ -78,6 +85,8 @@ class Provider(object):
             preload_content: Whether to preload the response content.
 
             as_json: Whether to deserialize the response body from JSON.
+
+            timeout: Timeout in seconds for API call
 
         Returns:
             Response body eventually deserialized.
@@ -98,11 +107,12 @@ class Provider(object):
         content = self._connection.put(path=path,
                                        headers=headers,
                                        data=data,
+                                       timeout=timeout or self.api_timeout,
                                        as_json=as_json,
                                        preload_content=preload_content)
         return content
 
-    def delete(self, path, *, preload_content=True, as_json=True):
+    def delete(self, path, *, preload_content=True, as_json=True, timeout=None):
         """Delete.
 
         Args:
@@ -111,6 +121,8 @@ class Provider(object):
             preload_content: Whether to preload the response content.
 
             as_json: Whether to deserialize the response body from JSON.
+
+            timeout: Timeout in seconds for API call
 
         Returns:
             Response body eventually deserialized.
@@ -121,6 +133,7 @@ class Provider(object):
         path = '{root}/{path}'.format(root=self._root_path, path=path)
         content = self._connection.delete(path=path,
                                           headers=headers,
+                                          timeout=timeout or self.api_timeout,
                                           preload_content=preload_content,
                                           as_json=True)
         return content
@@ -128,6 +141,7 @@ class Provider(object):
 
 class WithSearchRoute():
     _search_path = 'search'
+    timeout = DEFAULT_API_TIMEOUT
 
     def search(self, path, query):
         """Search objects.
@@ -139,6 +153,7 @@ class WithSearchRoute():
                                                                search_path=self._search_path),
                 headers={'Cache-Control': 'no-cache', 'Content-Type': 'application/json'},
                 data=json.dumps(query),
+                timeout=self.timeout,
                 as_json=True)
 
         return content
@@ -146,39 +161,49 @@ class WithSearchRoute():
 
 class AnnotationsAPI(Provider):
     _root_path = 'map-service/annotations'
+    api_timeout = DEFAULT_API_TIMEOUT  # todo: define from service
 
 
 class AuthAPI(WithSearchRoute, Provider):
     _root_path = 'dxauth'
+    api_timeout = DEFAULT_API_TIMEOUT  # todo: define from service
 
 
 class ProjectManagerAPI(WithSearchRoute, Provider):
     _root_path = 'dxpm'
+    api_timeout = DEFAULT_API_TIMEOUT  # todo: define from service
 
 
 class DataManagementAPI(Provider):
     _root_path = 'data-manager'
+    api_timeout = DEFAULT_API_TIMEOUT  # todo: define from service
 
 
 class UIServicesAPI(Provider):
     _root_path = 'uisrv'
+    api_timeout = DEFAULT_API_TIMEOUT  # todo: define from service
 
 
 class AnalyticsServiceAPI(Provider):
     _root_path = 'analytics-service'
+    api_timeout = DEFAULT_API_TIMEOUT  # todo: define from service
 
 
 class ExternalProviderServiceAPI(Provider):
     _root_path = 'external-providers-service'
+    api_timeout = DEFAULT_API_TIMEOUT  # todo: define from service
 
 
 class AssetManagementAPI(Provider):
     _root_path = 'dct-service/asset-management'
+    api_timeout = DEFAULT_API_TIMEOUT  # todo: define from service
 
 
 class CollectionTaskAPI(Provider):
     _root_path = 'dct-service/task'
+    api_timeout = DEFAULT_API_TIMEOUT  # todo: define from service
 
 
 class CollectionTaskManagementAPI(Provider):
     _root_path = 'dct-service/task-management'
+    api_timeout = DEFAULT_API_TIMEOUT  # todo: define from service